New/services/llm-router/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
httpx==0.26.0
anthropic==0.18.1
pyyaml==6.0.1
prometheus-client==0.19.0
